Integrate function export scripts into ExportFunctionInfoScript
- Consolidated ExportFunctionMetricsToJson, ExportFunctionMetricsToCsv, and ExportFunctionCallGraphToDot into ExportFunctionInfoScript.java
- Added format selection using GhidraScript.askChoices() as requested in code review
- Users can now choose between: JSON (Simple), JSON (Detailed metrics), CSV (Detailed metrics), and DOT (Function call graph)
- All export formats support optional selection/highlight filtering
- Maintains backward compatibility with existing simple JSON export functionality
